Implement interactive template selection
The phishingpage module was redesigned to be more OOP which would
make any modifications easier to implement. This patch will enable
wifiphisher to show all the available templates to the user and
ask for one to be chosen. It will also look for any custom
templates that the user already has.