Add proposed changes This patch will add all the proposed changes in #233. The fallowing summarize these changes:
The whole module is re-designed with OOP in mind. All the previous functions are converted to methods for different classes.

More generic methods are added to help reduce the required code to run Linux commands.

More custom errors are added and existing ones are updated with new messages. These messages should help users understand the error better.

The documentation style is changed to sphinx so it can be done automatically.