refactor(macmatcher): refactor macmatcher for better code readability
This patch will refactor macmatcher for better code readability. It also adds more comments and
documentation to the macmatcher module. A note worthy change is that get_vendor_name method no
longer returns False when it is unable to find a vendor. It now return "Unknown" to indicate a
unknown vendor.