fix(pywifiphisher):fix ouput window flickering
The original cause of this bug is the use of hard coded values for
row numbers as it would cause issues when the window size is less
than the row number. This patch will not fix the above bug however
it does ensure that the minimum number of rows are used to achieve
the output.