fix(recon):error happening on ap selection
This patch will fix the error in ap selection where it would cause a
TypeError in the program. The cause of this error is the ap name
when the ap in hidden. It only shows the number of bits in format
"\00". Example if the name of the ap is HELLO it would show it as
"\00\00\00\00\00".

issue #506