refactor(*): implement a new design for interfaces
Implement a new design for interfaces module where getter and setters
are removed. Some of the methods have been removed or replaced to
make using network manager class much easier. Error messages are
simplified as well. Most importantly a tests suite is written for
the module and the coverage is 100%.