Social media Login
Social media Oauth login fake popup (Mac OS, Windows,Linux, Android, IOS Friendly), Notice that the popup will be launched only in Mac OS and Windows. For the other OSs it will be redirected to a facebook login page (twitter login and google+ will be available in the next update).